Sync DefaultUserPath and DefaultSystemPath with _PATH_DEFPATH and _PATH_STDPATH from <paths.h>.
Fixes second half of PR xsrc/7638 from Rob Windsor (windsor@warthog.com),
but in a slightly different way.

The theory is that src/ and xsrc/ should agree on what a rational default path
is.  Note, however, that in addition to adding /usr/pkg paths to these paths,
this adds /usr/local paths to DefaultSystemPath.

Note also that ${X11BASE}/sbin is not added to DefaultSystemPath, as it does
not have a relocation macro (like $(BINDIR) to allow it to move with ${X11BASE}
(which NetBSD.cf otherwise obeys)