Add support for cats to X 4.2.0.
Note that some changes are generic arm changes, some aren't but are enclosed with #if defined(__arm__).  Currently the only arm port supported by X 4.2.0 is cats, so this is not a problem.  However shark support should be relatively easy based upon on the cats changes.

Ideally a few more things should be abstracted out, so that we can run the same X server on shark and cats, but currently my aim is to get X working just in time for NetBSD 1.6 release.

This has been tested on a s3virge card on my cats, I've also checked that i386 isn't broken by any of these changes.